Add options to offset search results, create tests
Search now has options to produce offset queries. This should make it easier
to do pagination like operations. Simply adding a `offset : $value` parameter
in the search options would do that. Unfortunately, pages still need to be
scraped for getting the IDs. Ideal way would be to just offset to closest
page and then start scraping as that would be quite a bit faster. That said,
this is a feature addition.

Two corresponding tests have been added too,

 - One which offsets queries by 10 and gets the next 10 results; check that
 only 10 results are present
 - Second which tests that these 10 offset results are contained in the
 30 results fetch from a basic search query. Check that a full fetch of 30
 resultls has all items that are fetched by the offset fetch.